biloCNN/requirements.txt
brotlipy           
certifi            
cffi               
charset-normalizer 
colorama           
cryptography       
idna               
mkl-fft            
mkl-random         
mkl-service        
numpy              
Pillow             
pip                
protobuf           
pycparser          
pyOpenSSL          
PySocks            
PyYAML             
requests           
setproctitle       
setuptools         
six                
tensorboardX       
torch              
torchaudio         
torchvision        
tqdm               
typing_extensions  
urllib3            
wheel              
opencv-python      
